+++ core/doc/markdown_files/index.md
@@ -46,7 +46,7 @@ A virtual key board event and modifier map recevied from the Platform layer to b
 processed with the state object for this Client application.
 - __Virtual Key:__
 A code based on the US English layout, with values matching the Windows
-virtual key codes. See `keyboardprocessor_vkeys.h` for definitions.
+virtual key codes. See `keyman_core_api_vkeys.h` for definitions.
 - __Modifier Key:__
 The set of Control, Shift, Alt, Caps Lock keys. On some platforms these may
 have other names (e.g. Alt is called Option on macOS); other platform-specific
