+++ core/include/keyman/keyman_core_api_consts.h
