+++ core/include/keyman/keyman_core_api_version.h.in
