+++ core/src/km_kbp_context_api.cpp
@@ -12,7 +12,7 @@
 #include <algorithm>
 #include <vector>
 
-#include <keyman/keyboardprocessor.h>
+#include <keyman/keyman_core_api.h>
 
 #include "context.hpp"
 #include "jsonpp.hpp"
