+++ core/src/km_kbp_options_api.cpp
@@ -11,7 +11,7 @@
 #include <cassert>
 #include <sstream>
 
-#include <keyman/keyboardprocessor.h>
+#include <keyman/keyman_core_api.h>
 #include "processor.hpp"
 
 #include "jsonpp.hpp"
