+++ core/src/kmx/kmx_capslock.cpp
@@ -2,7 +2,7 @@
   Copyright:        Copyright (C) 2003-2018 SIL International.
   Authors:          mcdurdin
 */
-#include <keyman/keyboardprocessor_consts.h>
+#include <keyman/keyman_core_api_consts.h.h>
 #include <kmx/kmx_processevent.h>
 
 using namespace km::kbp;
