+++ core/src/kmx/kmx_options.h
@@ -4,7 +4,7 @@
 #include <string>
 #include <vector>
 
-#include <keyman/keyboardprocessor.h>
+#include <keyman/keyman_core_api.h>
 #include "option.hpp"
 
 #include "kmx_base.h"
