+++ core/src/kmx/kmx_processevent.cpp
@@ -4,7 +4,7 @@
 */
 #include "kmx_processevent.h"
 #include "state.hpp"
-#include <keyman/keyboardprocessor_consts.h>
+#include <keyman/keyman_core_api_consts.h.h>
 
 using namespace km::kbp;
 using namespace kmx;
