+++ core/src/kmx/kmx_processevent.h
@@ -11,7 +11,7 @@
 #include <assert.h>
 #include <string>
 #include <string.h>
-#include <keyman/keyboardprocessor_bits.h>
+#include <keyman/keyman_core_api_bits.h>
 #include "debuglog.h"
 #include "kmx_base.h"
 #include "kmx_file.h"
