+++ core/src/kmx/kmx_processor.hpp
@@ -9,7 +9,7 @@
 #pragma once
 
 #include <string>
-#include <keyman/keyboardprocessor.h>
+#include <keyman/keyman_core_api.h>
 #include "kmx/kmx_processevent.h"
 #include "keyboard.hpp"
 #include "processor.hpp"
