+++ core/src/ldml/ldml_vkeys.hpp
@@ -13,7 +13,7 @@
 #include <utility>
 #include <vector>
 
-#include <keyman/keyboardprocessor.h>
+#include <keyman/keyman_core_api.h>
 
 namespace km {
 namespace kbp {
