+++ core/src/option.hpp
@@ -12,7 +12,7 @@
 
 #include <string>
 
-#include <keyman/keyboardprocessor.h>
+#include <keyman/keyman_core_api.h>
 
 // Forward declarations
 class json;
