+++ core/tests/unit/kmnkbd/context_api.cpp
@@ -10,7 +10,7 @@
                                     mutation functions.
 */
 #include <string>
-#include <keyman/keyboardprocessor.h>
+#include <keyman/keyman_core_api.h>
 
 #include "context.hpp"
 #include "utfcodec.hpp"
