+++ core/tests/unit/kmnkbd/debug_api.cpp
@@ -10,7 +10,7 @@
 #include <algorithm>
 #include <iterator>
 #include <string>
-#include <keyman/keyboardprocessor.h>
+#include <keyman/keyman_core_api.h>
 #include "path.hpp"
 #include "state.hpp"
 #include "kmx/kmx_base.h"
@@ -462,14 +462,14 @@ int error_args() {
 
 int print_sizeof() {
   std::cout << std::endl;
-  std::cout << "keyboardprocessor.h:" << std::endl;
+  std::cout << "keyman_core_api.h:" << std::endl;
   std::cout << "sizeof(km_kbp_context_item): " << sizeof(km_kbp_context_item) << std::endl;
   std::cout << "sizeof(km_kbp_action_item): " << sizeof(km_kbp_action_item) << std::endl;
   std::cout << "sizeof(km_kbp_option_item): " << sizeof(km_kbp_option_item) << std::endl;
   std::cout << "sizeof(km_kbp_keyboard_attrs): " << sizeof(km_kbp_keyboard_attrs) << std::endl;
   std::cout << "sizeof(km_kbp_attr): " << sizeof(km_kbp_attr) << std::endl;
   std::cout << std::endl;
-  std::cout << "keyboardprocessor_debug.h:" << std::endl;
+  std::cout << "keyman_core_api_debug.h.h:" << std::endl;
   std::cout << "sizeof(km_kbp_state_debug_item): " << sizeof(km_kbp_state_debug_item) << std::endl;
   std::cout << "sizeof(km_kbp_state_debug_key_info): " << sizeof(km_kbp_state_debug_key_info) << std::endl;
   std::cout << "sizeof(km_kbp_state_debug_kmx_info): " << sizeof(km_kbp_state_debug_kmx_info) << std::endl;
