+++ core/tests/unit/kmnkbd/debug_items.hpp
@@ -1,6 +1,6 @@
 #include <iostream>
-#include <keyman/keyboardprocessor.h>
-#include <keyman/keyboardprocessor_debug.h>
+#include <keyman/keyman_core_api.h>
+#include <keyman/keyman_core_api_debug.h.h>
 #include "kmx/kmx_base.h"
 #include "kmx/kmx_xstring.h"
