+++ core/tests/unit/ldml/ldml_test_utils.hpp
@@ -13,7 +13,7 @@
 #include <kmx/kmx_processevent.h> // for char to vk mapping tables
 #include <kmx/kmx_xstring.h> // for surrogate pair macros
 #include <kmx/kmx_plus.h>
-#include <keyman/keyboardprocessor.h>
+#include <keyman/keyman_core_api.h>
 
 namespace km {
 namespace tests {
