+++ developer/src/tike/main/Keyman.System.KeymanCoreDebug.pas
@@ -1,7 +1,7 @@
 {
  * Keyman is copyright (C) SIL International. MIT License.
  *
- * Debug interface for Keyman Core, matches keyboardprocessor_debug.h
+ * Debug interface for Keyman Core, matches keyman_core_api_debug.h.h
 }
 unit Keyman.System.KeymanCoreDebug;
 
@@ -181,7 +181,7 @@ initialization
   // Assertions generated by debug_api.cpp for win32 x86.
   // Usage: debug-api.exe --print-sizeof
 
-  //keyboardprocessor.h:
+  //keyman_core_api.h:
   Assert(sizeof(km_kbp_context_item) = 8);
   Assert(sizeof(km_kbp_context_item) = 8);
   Assert(sizeof(km_kbp_action_item) = 12);
@@ -189,7 +189,7 @@ initialization
   Assert(sizeof(km_kbp_keyboard_attrs) = 16);
   Assert(sizeof(km_kbp_attr) = 16);
 
-  //keyboardprocessor_debug.h:
+  //keyman_core_api_debug.h.h:
   Assert(sizeof(km_kbp_state_debug_item) = 432);
   Assert(sizeof(km_kbp_state_debug_key_info) = 6);
   Assert(sizeof(km_kbp_state_debug_kmx_info) = 416);
